changed that now most of the services using GenericService to setFields soon UserService will use this service and the service will be added the find function
