--- DebriefRepository/src/main/java/org/example/debriefrepository/repository/ContentItemRepository.java
@@ -1,7 +1,0 @@
-package org.example.debriefrepository.repository;
-
-import org.example.debriefrepository.entity.ContentItem;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ContentItemRepository extends JpaRepository<ContentItem, String> {
-}
