+++ DebriefRepository/src/main/java/org/example/debriefrepository/service/CommentService.java
@@ -8,20 +8,32 @@ import org.example.debriefrepository.repository.ParagraphRepository;
 import org.example.debriefrepository.types.content.CommentInput;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.util.ArrayList;
+import java.util.List;
 
 @Service
 @RequiredArgsConstructor
-public class CommentService extends OrderedItemService {
+public class CommentService {
 
     private final CommentRepository commentRepository;
+
     private final ParagraphRepository paragraphRepository;
+
+    @Autowired
+    private final GenericService<Comment, CommentInput> genericService;
+
     private static final Logger logger = LoggerFactory.getLogger(CommentService.class);
 
     public Comment createComment(CommentInput input, String paragraphId) {
         Comment comment = new Comment();
         try {
-            comment = setFields(comment, input);
+            List<String> skippedFields = new ArrayList<>();
+            skippedFields.add("id");
+            skippedFields.add("paragraph");
+            comment = setFields(comment, input, skippedFields);
             comment.setParagraph(paragraphRepository.findById(paragraphId).
                     orElseThrow(() -> new IllegalArgumentException("Invalid paragraph id: " + paragraphId)));
             return commentRepository.save(comment);
@@ -33,21 +45,16 @@ public class CommentService extends OrderedItemService {
 
     public Comment updateComment(CommentInput input) {
         String id = input.getId();
+        List<String> skippedFields = new ArrayList<>();
+        skippedFields.add("id");
         Comment comment = commentRepository.findById(id)
                 .orElseThrow(() -> new IllegalArgumentException("Comment not found"));
-        comment = setFields(comment, input);
+        comment = setFields(comment, input, skippedFields);
         return commentRepository.save(comment);
     }
 
-    protected Comment setFields(Comment comment, CommentInput input) {
-        comment = (Comment) super.setFields(comment, input);
-
-        if (input.getBullet() == null) {
-            throw new IllegalArgumentException("Bullet cannot be null");
-        }
-
-        comment.setBullet(input.getBullet());
-        return comment;
+    protected Comment setFields(Comment comment, CommentInput input, List<String> skippedFields) {
+        return genericService.setFieldsGeneric(comment, input, null, skippedFields);
     }
 
 }
